+++ packages/components/src/index.ts
@@ -13,7 +13,6 @@ export * from './Checkbox';
 export * from './Columns';
 export * from './Dialog';
 export * from './Divider';
-export * from './Field';
 export * from './VisuallyHidden';
 export * from './Image';
 export * from './Inline';
@@ -30,6 +29,7 @@ export * from './Switch';
 export * from './Stack';
 export * from './Text';
 export * from './Textarea';
+export * from './TextField';
 export * from './Tiles';
 export * from './Tooltip';
 export * from './Input';
